requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.7.4
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.0
daphne==4.1.2
Django==4.1
django-cors-headers==4.4.0
django-cursor-pagination==0.3.0
django-filter==23.1
djangorestframework==3.14.0
dnspython==2.6.1
et_xmlfile==2.0.0
filelock==3.16.1
fsspec==2024.9.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
itypes==1.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pillow==10.4.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-whois==0.9.4
pytz==2024.1
requests==2.32.3
requests-file==2.1.0
scipy==1.14.1
service-identity==24.1.0
setuptools==71.0.3
six==1.16.0
sqlparse==0.5.1
sympy==1.13.3
tldextract==5.1.2
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.2
watchdog==4.0.1
zope.interface==6.4.post2
